Don't swallow exceptions in GlueJob class
Exceptions are a value source of information that can be used
to debug problems. By silencing them, the end user (the writer of
the Glue Job) will not be able to understand if something went wrong
or the job succeeded and will not know what was the cause of the
problem: Was it a connectivity issue? was it a credentials problems?
was it a quota problem? who knows, it's hard to tell without
a stacktrace.

I've added two exception classes:
- `JobMisconfigured` for when the user didn't set the `job_name`
- `JobNotStarted` for when the user tried to get the status of the job
  but the job doesn't have a `run_id` (possibly because the user
  didn't start the job yet

This is just a starting point, we may want to handle some exceptions
and add more exceptions classes but it's a good start I think.

NOTE: This may potentially break the client code.